3-redirection.sh

+++ 0x0C-web_server/3-redirection.sh
@@ -1,22 +1,25 @@
 #!/usr/bin/env bash
 
 # Install nginx web server and start it
-apt-get -y update
-apt-get -y install nginx
-service nginx start
+# Install Nginx if it is not already installed
+if ! dpkg -l | grep -q nginx; then
+    apt-get -y update
+    apt-get -y install nginx
+    service nginx start
+fi
 
 # Firewall setting: Allow Nginx to serve on HTTP
 ufw allow 'Nginx HTTP'
 
 # Override the default index file served as home page
-echo "Hello World!" >> /var/www/html/index.html
+[ ! -e /var/www/html/index.nginx-debian-bk.html ] && cp /var/www/html/index.nginx-debian.html /var/www/html/index.nginx-debian-bk.html
+echo "Hello World!" | sudo tee /var/www/html/index.html
 
-#Give the user  ownership to website files
-sudo chown -R "$USER":"$USER" /var/www/html
-sudo chmod -R 755 /var/www
+CONFIG_FILE="/etc/nginx/sites-available/default"
 
-# Redirect to /redirect_me to a youtube video
-sed -i '47i\\tlocation /redirect_me { return 301 https://www.youtube.com/watch?v=QH2-TGUlwu4; }' /etc/nginx/sites-available/default
+if ! grep -q "location /redirect_me" "$CONFIG_FILE"; then
+    sed -i "/server_name _;/a location /redirect_me {\n return 301 https://www.youtube.com/watch?v=QH2-TGUlwu4;\n}" "$CONFIG_FILE"
+fi
 
 # Restart nginx after new configurations
 service nginx restart
